Refactor selenium test
