+++ pom.xml
@@ -6,6 +6,7 @@
 
     <groupId>com.selenium</groupId>
     <artifactId>Selenium</artifactId>
+    <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
