+++ src/main/java/selenium/pages/GmailUserPage.java
@@ -1,4 +1,4 @@
-package selenium.pages.gmail;
+package selenium.pages;
 
 
 import org.openqa.selenium.WebDriver;
@@ -9,8 +9,8 @@ public class GmailUserPage {
     private WebDriver webDriver;
 
 
-
     public GmailUserPage(WebDriver webDriver) {
+
         this.webDriver = webDriver;
     }
