+++ src/main/java/selenium/pages/StylusProductPage.java
@@ -1,4 +1,4 @@
-package selenium.pages.stylus;
+package selenium.pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
