+++ src/test/java/com/selenium/LoginToGmailTest.java
@@ -1,30 +1,26 @@
-package com.selenium.gmail;
+package com.selenium;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import selenium.core.WebDriverTestBase;
-import selenium.pages.gmail.GmailLoginPage;
-import selenium.pages.gmail.GmailUserPage;
+import selenium.pages.GmailLoginPage;
+import selenium.pages.GmailUserPage;
 
 
 public class LoginToGmailTest extends WebDriverTestBase {
     private String urlGmail = "http://gmail.com";
     private String emailForTest = "jacksonford292@gmail.com";
     private String passwordForTest = "SuperSecret123";
-    private String verifyTitle = "Gmail";
+
 
 
     @Test
     public void gmailLogIn() {
-
+        driver.get(urlGmail);
         GmailLoginPage gmailLoginPage = new GmailLoginPage(driver);
-        gmailLoginPage.userEmailPassword(emailForTest, passwordForTest,urlGmail);
+        Assert.assertTrue(gmailLoginPage.userEmailPassword(emailForTest, passwordForTest));
     }
 
-    @Test()
-    public void verifyGmail() {
-        Assert.assertEquals(driver.getTitle(), verifyTitle);
-    }
 
     @Test
     public void verifyLoginGmail() {
